+++ dpws/src/main/java/org/somda/sdc/dpws/http/apache/CommunicationLogHttpResponseInterceptor.java
@@ -4,10 +4,11 @@ import org.apache.http.HttpEntity;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpResponseInterceptor;
+import org.apache.http.conn.ManagedHttpClientConnection;
 import org.apache.http.protocol.HttpContext;
 import org.apache.http.protocol.HttpCoreContext;
-import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.somda.sdc.common.logging.InstanceLogger;
 import org.somda.sdc.dpws.CommunicationLog;
 import org.somda.sdc.dpws.http.apache.helper.ApacheClientHelper;
@@ -15,9 +16,15 @@ import org.somda.sdc.dpws.soap.CommunicationContext;
 import org.somda.sdc.dpws.soap.HttpApplicationInfo;
 import org.somda.sdc.dpws.soap.TransportInfo;
 
+import javax.net.ssl.SSLPeerUnverifiedException;
+import javax.net.ssl.SSLSession;
 import java.io.OutputStream;
-import java.util.Collections;
+import java.security.cert.X509Certificate;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * Response interceptor which writes the incoming response message and headers into the {@linkplain CommunicationLog}.
@@ -51,6 +58,28 @@ public class CommunicationLogHttpResponseInterceptor implements HttpResponseInte
                 null
         );
 
+        final List<X509Certificate> x509certificates = new ArrayList<>();
+        ManagedHttpClientConnection routedConnection = null;
+        try {
+            routedConnection = (ManagedHttpClientConnection) context.getAttribute(
+                    HttpCoreContext.HTTP_CONNECTION);
+        } catch (ClassCastException e) {
+            LOG.error("Error retrieving managed http client connection: {}", e.getMessage(), e);
+        }
+        if (routedConnection != null && routedConnection.isOpen()) {
+            SSLSession sslSession = routedConnection.getSSLSession();
+            try {
+                if (sslSession != null
+                        && sslSession.getPeerCertificates() != null) {
+                    x509certificates.addAll(Arrays.stream(sslSession.getPeerCertificates())
+                            .filter(certificate -> certificate instanceof X509Certificate)
+                            .map(certificate -> (X509Certificate) certificate).collect(Collectors.toList()));
+                }
+            } catch (SSLPeerUnverifiedException e) {
+                LOG.error("Error retrieving peer certificates: {} ", e.getMessage(), e);
+            }
+        }
+
         // collect information for TransportInfo
         var requestTransportInfo = new TransportInfo(
                 target.getSchemeName(),
@@ -58,7 +87,7 @@ public class CommunicationLogHttpResponseInterceptor implements HttpResponseInte
                 null,
                 target.getHostName(),
                 target.getPort(),
-                Collections.emptyList()
+                x509certificates
         );
 
         var requestCommContext = new CommunicationContext(requestHttpApplicationInfo, requestTransportInfo);